+++ bindings/wasm/src/account/wasm_auto_save.rs
