+++ bindings/wasm/src/account/wasm_identity_setup.rs
@@ -1,7 +1,7 @@
 // Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::account::method_secret::WasmMethodSecret;
+use crate::account::wasm_method_secret::WasmMethodSecret;
 use crate::crypto::KeyType;
 use wasm_bindgen::prelude::*;
