+++ bindings/wasm/src/account/wasm_method_secret.rs
